Added pathfinding code 12/27/21
Added code to find the path between the entrance and the exit. This is done to create a set of rooms in which the player can traverse through and exit the floor. I also added a function with the ability to add many rooms from a list to visually see the path.