Finished Branch Function 1/1/22
(Hopefully) Finished with branch functions to create branches. To create branches, though, functions outside the class will be required so I will probably have to code that into the game part of the program. I also added a checkPath function since it's the easiest way to check a list of coordinates without having to make sure that the original function outputs some proper coordinates.